third_party/WebKit/LayoutTests/bluetooth/server/device-same-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getDiscoveredHealthThermometerDevice()
    .then(([device, fake_peripheral]) => {
      return fake_peripheral.setNextGATTConnectionResponse({code: HCI_SUCCESS})
        .then(() => device.gatt.connect());
    })
    .then(gatt => {
      assert_equals(gatt.device, gatt.device);
    });
}, "[SameObject] test for BluetoothRemoteGATTServer's device.");
</script>
